First batch of changes prior to testing
