+++ exporter/SynthesisFusionAddin/src/UI/colors.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class Colors(str, Enum):
+    background = "#d9d9d9"
